refactor: move explainSolTransaction to lib/explainTransactionWasm.ts
Follow the *Wasm.ts naming convention used by abstract-utxo
(explainPsbtWasm.ts, signPsbtWasm.ts). Moves the standalone WASM
explain adapter out of sol.ts into its own module with a dedicated
interface, barrel-exported from lib/index.ts.

BTC-0

TICKET: BTC-0
